ctype functions take ints
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/trunk@1189 96dfbd92-d197-e211-9ca9-00110a534b34

+++ as6502/linectl.c
@@ -102,7 +102,7 @@ char *trimheadtospc(const char *str, size_t len) {
 }
 
 char *strnspc(const char *str, size_t len) {
-	while (len && !isspace(*str)) {
+	while (len && !isspace((int)*str)) {
 		str++;
 		len--;
 	}
@@ -111,7 +111,7 @@ char *strnspc(const char *str, size_t len) {
 }
 
 char *strnpc(const char *str, size_t len) {
-	while (len && isspace(*str)) {
+	while (len && isspace((int)*str)) {
 		str++;
 		len--;
 	}
